interpreter: Split base objects and helpers from interpreterbase.py
